requirements.txt
streamlit
requests
psycopg2-binary
pgvector
datasets
pillow
python-dotenv
huggingface-hub
transformers>=4.49.0
torch
sentencepiece
protobuf
